@@ -48,10 +48,10 @@
     "@nuxtjs/stylelint-module": "^4.1.0",
     "@types/uuid": "^8.3.4",
     "apollo-boost": "^0.4.9",
-    "eslint": "^8.18.0",
+    "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-nuxt": "^3.2.0",
-    "eslint-plugin-prettier": "^4.1.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-vue": "^9.1.1",
     "node-fetch": "^3.2.6",
     "postcss-html": "^1.4.1",
