@@ -50,11 +50,12 @@
 </template>
 
 <script>
+import { defineComponent } from '@nuxtjs/composition-api'
 import { debounce } from 'throttle-debounce'
 import MainNavigationItem from '~/components/Menu/MainNavigation/MainNavigationItem.vue'
 import pages from '~/data/menu'
 
-export default {
+export default defineComponent({
   components: {
     MainNavigationItem,
   },
@@ -115,7 +116,7 @@ export default {
       return menu.querySelector('.menu-item-page .nuxt-link-active')
     },
   },
-}
+})
 </script>
 
 <style lang="postcss" module>
