@@ -1,43 +1,43 @@
 <template>
   <div v-if="cart">
     <h2>Jouw bestelling</h2>
-    <table :class="$style.list">
+    <table class="list">
       <thead>
         <tr>
           <th>Product</th>
-          <th :class="$style.price">Subtotaal</th>
+          <th class="price">Subtotaal</th>
         </tr>
       </thead>
       <tr v-for="item in cart.contents.nodes" :key="item.key">
         <td>
-          <div :class="$style.title">
-            <span :class="$style.quantity">{{ item.quantity }}x</span>
+          <div class="title">
+            <span class="quantity">{{ item.quantity }}x</span>
             {{ item.product.node.name }}
             {{ item.product.node.name }}
           </div>
         </td>
-        <td :class="$style.price">
+        <td class="price">
           {{ item.total }}
         </td>
       </tr>
       <tfoot>
         <tr>
-          <td :class="$style.label">Subtotaal</td>
-          <td :class="$style.price">{{ cart.subtotal }}</td>
+          <td class="label">Subtotaal</td>
+          <td class="price">{{ cart.subtotal }}</td>
         </tr>
         <template v-if="cart.appliedCoupons">
           <tr v-for="coupon in cart.appliedCoupons" :key="coupon.code">
-            <td :class="$style.label">Kortingsbon: {{ coupon.code }}</td>
-            <td :class="$style.price">- {{ coupon.discountAmount }}</td>
+            <td class="label">Kortingsbon: {{ coupon.code }}</td>
+            <td class="price">- {{ coupon.discountAmount }}</td>
           </tr>
         </template>
         <tr>
-          <td :class="$style.label">Verzending</td>
-          <td :class="$style.price">{{ cart.shippingTotal }}</td>
+          <td class="label">Verzending</td>
+          <td class="price">{{ cart.shippingTotal }}</td>
         </tr>
         <tr>
-          <td :class="$style.total">Totaal</td>
-          <td :class="[$style.price, $style.total]">{{ cart.total }}</td>
+          <td class="total">Totaal</td>
+          <td class="price total">{{ cart.total }}</td>
         </tr>
       </tfoot>
     </table>
@@ -58,7 +58,7 @@ export default {
 }
 </script>
 
-<style lang="postcss" module>
+<style lang="postcss" scoped>
 .price {
   font-variant-numeric: tabular-nums;
   text-align: right;
