@@ -59,16 +59,16 @@ export default {
       websocketsOnly: false,
     })
 
-    return []
+    // return []
 
-    // const posters = await fetchPosters(client)
-    // await pauseFetching()
-    // const pages = await fetchPages(client)
-    // await pauseFetching()
+    const posters = await fetchPosters(client)
+    await pauseFetching()
+    const pages = await fetchPages(client)
+    await pauseFetching()
     // const productCategories = await fetchProductCategories(client)
     // await pauseFetching()
-    // const posts = await fetchPosts(client)
+    const posts = await fetchPosts(client)
 
-    // return [...pages, ...productCategories, ...posts, ...posters]
+    return [...pages, ...posts, ...posters]
   },
 }
