@@ -1,5 +1,5 @@
 export const FETCH_TIMEOUT = 2000
-export const TOTAL_PAGES = 10
-export const TOTAL_POSTS = 10
+export const TOTAL_PAGES = 50
+export const TOTAL_POSTS = 50
 export const TOTAL_POSTERS = 25
 export const TOTAL_PRODUCT_CATEGORIES = 10
