@@ -24,7 +24,7 @@
   </page-with-sidebar>
 </template>
 
-<script>
+<script lang="ts">
 import { defineComponent, useContext } from '@nuxtjs/composition-api'
 
 export default defineComponent({
