@@ -4,13 +4,15 @@
   </shop-wrapper>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from '@nuxtjs/composition-api'
+
+export default defineComponent({
   middleware: ['auth'],
   nuxtI18n: {
     paths: {
       nl: '/winkeltje/registreren',
     },
   },
-}
+})
 </script>
